+++ doc/changelog.rst
@@ -93,6 +93,7 @@ Addons
 Front
 ~~~~~
 
+- Fix bug with multiple service checkout phases
 - Update UI for package products
 - Add SHOOP_FRONT_ADDRESS_FIELD_PROPERTIES setting
 - Support also django-registration-redux 1.4